feat(ui): add skeletal create alert form and routing
